+++ BRDFs/BRDF.h
@@ -28,7 +28,7 @@ class BRDF {
 		sample_f(const ShadeRec& sr, const Vector3D& wo, Vector3D& wi) const;
 		
 		virtual RGBColor
-		sample_f(const ShadeRec& sr, const Vector3D& wo, Vector3D& wi, float& pdf) const;
+		sample_f(const ShadeRec& sr, const Vector3D& wo, Vector3D& wi, double &pdf) const;
 		
 		virtual RGBColor
 		rho(const ShadeRec& sr, const Vector3D& wo) const;
